# Cherry-pick #20717 to 7.x: Improve registry file migration performance